requirements.txt
aioredis==1.3.1
amqp==2.5.2
APScheduler==3.6.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.2.1
Automat==0.8.0
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
dj-database-url==0.5.0
Django==2.2.5
django-bootstrap3==12.0.3
django-crispy-forms==1.7.2
django-filter==2.2.0
django-filters==0.2.1
django-heroku==0.3.1
django-tables2==2.2.1
djangorestframework==3.10.3
gunicorn==19.9.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
kombu==4.6.7
msgpack==0.6.2
Pillow==7.0.0
psycopg2==2.8.3
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.0
pyOpenSSL==19.1.0
pytz==2019.2
qrcode==6.1
redis==3.3.11
reportlab==3.5.28
requests==2.22.0
rq==1.1.0
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
tzlocal==2.0.0
urllib3==1.25.7
vine==1.3.0
whitenoise==4.1.4
zope.interface==4.7.1
